requirements.txt
python-telegram-bot==20.0
mercadopago==2.1.0
Flask==2.0.1